third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/navigate-in-transition-finished_currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering when navigate() is called inside the transition.finished promise handler
  assert_array_equals: expected property 3 to be "navigatesuccess" but got "committed fulfilled 1" (expected array ["navigate", "currententrychange", "handler run", "navigatesuccess", "committed fulfilled 1", "transition.committed fulfilled 1", "promise microtask", "finished fulfilled 1", "transition.finished fulfilled", "navigate", "currententrychange", "handler run", "navigatesuccess", "committed fulfilled 2", "transition.committed fulfilled 2", "finished fulfilled 2", "transition.finished fulfilled"] got ["navigate", "currententrychange", "handler run", "committed fulfilled 1", "transition.committed fulfilled 1", "promise microtask", "navigatesuccess", "finished fulfilled 1", "transition.finished fulfilled", "navigate", "currententrychange", "handler run", "committed fulfilled 2", "transition.committed fulfilled 2", "navigatesuccess", "finished fulfilled 2", "transition.finished fulfilled"])
Harness: the test ran to completion.
